Data Analysis Project
Happiness score is determined by 6 indicators based on a global wellness institute website. These indicators are income, freedom, trust in government, healthy life expectancy, social support from family and friends, and generosity.

Also, I went to research on how Finland was able to stay at the top for many years and according to Helinksi Times, it did not mention anything about those indicators but said that it was because of the low crime rates and good standard of living. So, I wanted to see what was Singapore's take on these factors. I was only able to find data on crime cases and and the unemployemnt levels in Singapore. Therefore, since Finland was able to stay at the top due to low crime cses and low unemployment, I wanted to see of there is a correlation between these two factors on the happiness score of Singapore from 2016 to 2021..